requirements.txt
torch>=1.13.0
torchvision>=0.14.0
opencv-python>=4.7.0.72
numpy>=1.23.5
Pillow>=9.3.0
ultralytics>=8.0.20
tensorflow>=2.11.0
keras-preprocessing>=1.1.2
